:bug: fix: User관련 메서드 수정
