+++ server/src/main/java/server/team33/audit/Auditable.java
@@ -15,15 +15,11 @@ import java.time.ZonedDateTime;
 public abstract class Auditable {
     private ZonedDateTime createdAt;
     private ZonedDateTime updatedAt;
-
-
     @PrePersist
     public void prePersist(){
         this.createdAt = ZonedDateTime.now();
         this.updatedAt = ZonedDateTime.now();
-
     }
-
     @PreUpdate
     public void preUpdate(){
         this.updatedAt = ZonedDateTime.now();
