+++ server/src/main/java/server/team33/login/dto/LoginDto.java
@@ -1,0 +1,15 @@
+package server.team33.login.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginDto {
+    private String username;
+    private String password;
+}
