+++ server/src/main/java/server/team33/login/oauth/provider/OAuth2UserInfo.java
@@ -1,0 +1,8 @@
+package server.team33.login.oauth.provider;
+
+public interface OAuth2UserInfo {
+    String getProviderId();
+    String getProvider();
+    String getEmail();
+    String getName();
+}
