+++ server/src/main/java/server/team33/user/dto/UserSignUpDto.java
@@ -1,5 +1,6 @@
 package server.team33.user.dto;
 
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -7,6 +8,7 @@ import javax.validation.constraints.NotBlank;
 
 @Getter
 @Setter
+@Builder
 public class UserSignUpDto {
     @NotBlank
     private String email;
@@ -19,6 +21,6 @@ public class UserSignUpDto {
     @NotBlank
     private String realName;
     @NotBlank
-    private String phoneNumber;
+    private String phone;
 
 }
