--- server/src/main/java/server/team33/user/login/dto/LoginDto.java
@@ -1,11 +1,0 @@
-package server.team33.user.login.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class LoginDto {
-    private String username;
-    private String password;
-}
