--- server/src/main/java/server/team33/user/login/filter/JwtLoginFilter.java
@@ -1,48 +1,0 @@
-package server.team33.user.login.filter;
-
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import server.team33.user.login.dto.LoginDto;
-import server.team33.user.repository.UserRepository;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@RequiredArgsConstructor
-@Slf4j
-public class JwtLoginFilter extends UsernamePasswordAuthenticationFilter {
-
-
-    private final AuthenticationManager authenticationManager;
-    private final UserRepository userRepository;
-
-
-    @Override
-    @SneakyThrows
-    public Authentication attemptAuthentication( HttpServletRequest request, HttpServletResponse response ) throws AuthenticationException{
-        ObjectMapper objectMapper = new ObjectMapper();
-        LoginDto loginDto = objectMapper.readValue(request.getInputStream(), LoginDto.class);
-
-        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(loginDto.getUsername(), loginDto.getPassword());
-        log.info("authentication = {}", authenticationToken);
-
-        return authenticationManager.authenticate(authenticationToken);
-    }
-
-    @Override
-    protected void successfulAuthentication( HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication authResult ) throws IOException, ServletException{
-        logger.info("successful Authentication");
-        this.getSuccessHandler().onAuthenticationSuccess(request, response, authResult);
-    }
-}
