--- server/src/main/java/server/team33/user/login/handler/UserAuthFailureHandler.java
@@ -1,23 +1,0 @@
-package server.team33.user.login.handler;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-@Slf4j
-public class UserAuthFailureHandler implements AuthenticationFailureHandler {
-
-    @Override
-    public void onAuthenticationFailure( HttpServletRequest request, HttpServletResponse response, AuthenticationException exception ) throws IOException, ServletException{
-        log.error("로그인 실패");
-        UserAuthenticationEntryPoint.errorToJson(response, HttpStatus.UNAUTHORIZED);
-    }
-}
