--- server/src/main/java/server/team33/user/login/oauth/provider/OAuth2UserInfo.java
@@ -1,8 +1,0 @@
-package server.team33.user.login.oauth.provider;
-
-public interface OAuth2UserInfo {
-    String getProviderId();
-    String getProvider();
-    String getEmail();
-    String getName();
-}
