+++ server/src/main/java/server/team33/user/repository/UserRepository.java
@@ -9,6 +9,6 @@ public interface UserRepository extends JpaRepository<User,Long> {
     Optional<User> findByEmail( String email);
     Optional<User> findByOauthId(String oauthId);
     Optional<User> findByDisplayName( String displayName);
-    Optional<User> findByPhoneNumber( String phoneNumber);
+    Optional<User> findByPhone( String phoneNumber);
 
 }
